Restructuring and fixing bug
