+++ src/config.json
@@ -1,0 +1,3 @@
+{
+    "name": "Dynamic Products"
+}
