+++ src/start.js
@@ -1,6 +1,18 @@
 let _modPath;
 
-const modName = "Dynamic Products";
+const config = require('./config.json');
+
+const {
+    registerCompetitor,
+    registerFeature,
+    registerFramework,
+    registerProduct,
+} = require('./registration');
+
+const {
+    getFeatures,
+    initLanguage,
+} = require('./helpers');
 
 const presets = [
     require('./presets/search-engine.json'),
@@ -10,30 +22,12 @@ const presets = [
     require('./presets/travel-planning.json'),
 ];
 
-const languages = {
-    en: require('./lang/en.json'),
-    de: require('./lang/de.json'),
-};
-
-function initLanguage() {
-    console.log('Loading languages...');
-    Object.entries(languages).forEach(entry => {
-        const [lang, data] = entry;
-        console.log('Loading', lang);
-        
-        Object.entries(data.strings).forEach(string => {
-            Modding.addTranslation(string[0], {[lang]: string[1]});
-        });
-    });
-    console.log('Loaded languages!');
-}
-
 exports.initialize = (modPath) => {
     _modPath = modPath;
 
     Modding.setMenuItem({
         name: 'dynamic',
-        tooltip: modName,
+        tooltip: config.name,
         tooltipPosition: 'top',
         faIcon: 'fa-cubes',
         badgeCount: 0,
@@ -148,10 +142,6 @@ exports.initialize = (modPath) => {
                 };
 
                 registerCompetitor(newCompetitor);
-                if (!GetRootScope().settings[modName].competitors) {
-                    GetRootScope().settings[modName].competitors = [];
-                }
-                GetRootScope().settings[modName].competitors.push(newCompetitor);
 
                 this.reset();
             };
@@ -220,10 +210,6 @@ exports.initialize = (modPath) => {
                 console.log(newFeature);
 
                 registerFeature(newFeature);
-                if (!GetRootScope().settings[modName].features) {
-                    GetRootScope().settings[modName].features = [];
-                }
-                GetRootScope().settings[modName].features.push(newFeature);
 
                 // Reset inputs
                 this.reset();
@@ -244,11 +230,7 @@ exports.initialize = (modPath) => {
                 };
                 console.log('New Framework', newFramework);
 
-                registerFramework(newFramework);
-                if(!GetRootScope().settings[modName].frameworks) {
-                    GetRootScope().settings[modName].frameworks = [];
-                }
-                GetRootScope().settings[modName].frameworks.push(newFramework);
+                registerFramework(newFramework, _modPath);
 
                 // Reset input fields
                 this.reset();
@@ -301,10 +283,6 @@ exports.initialize = (modPath) => {
                 console.log(newProduct);
 
                 registerProduct(newProduct);
-                if (!GetRootScope().settings[modName].products) {
-                    GetRootScope().settings[modName].products = [];
-                }
-                GetRootScope().settings[modName].products.push(newProduct);
 
                 // Reset inputs
                 this.reset();
@@ -322,7 +300,7 @@ exports.initialize = (modPath) => {
                     preset.features.forEach(feature => registerFeature(feature));
                 }
                 if(preset.frameworks) {
-                    preset.frameworks.forEach(framework => registerFramework(framework));
+                    preset.frameworks.forEach(framework => registerFramework(framework, _modPath));
                 }
                 if (preset.products) {
                     preset.products.forEach(product => registerProduct(product));
@@ -364,165 +342,30 @@ exports.initialize = (modPath) => {
 exports.onLoadGame = settings => {
     initLanguage();
 
-    if (!settings[modName]) {
-        settings[modName] = {
+    if (!settings[config.name]) {
+        settings[config.name] = {
             features: [],
             frameworks: [],
             products: [],
         };
     }
-    if(!settings[modName].features) {
-        settings[modName].features = [];
+    if(!settings[config.name].features) {
+        settings[config.name].features = [];
     }
-    if(!settings[modName].frameworks) {
-        settings[modName].frameworks = [];
+    if(!settings[config.name].frameworks) {
+        settings[config.name].frameworks = [];
     }
-    if(!settings[modName].products) {
-        settings[modName].products = [];
+    if(!settings[config.name].products) {
+        settings[config.name].products = [];
     }
 
-    settings[modName].features.forEach(feature => {
-        console.log('Registering feature', feature);
-        registerFeature(feature);
+    settings[config.name].features.forEach(feature => {
+        registerFeature(feature, false);
     });
-    settings[modName].frameworks.forEach(framework => {
-        console.log('Registering framework', framework);
-        registerFramework(framework);
+    settings[config.name].frameworks.forEach(framework => {
+        registerFramework(framework, _modPath, false);
     });
-    settings[modName].products.forEach(product => {
-        console.log('Registering product', product);
-        registerProduct(product);
+    settings[config.name].products.forEach(product => {
+        registerProduct(product, false);
     });
 };
-
-function getFeatures() {
-    const features = {};
-    Object.keys(FeatureNames).forEach(name => {
-        features[name] = false;
-    });
-    return features;
-}
-
-function getInternalName(name) {
-    return _.startCase(_.toLower(name)).replace(/ /g, '');
-}
-
-function registerCompetitor(competitor) {
-    if(GetRootScope().settings.competitorProducts.find(product => product.name == competitor.name)) {
-        console.log('Skipping competitor', competitor, 'as it already exists');
-        return;
-    }
-
-    const internalName = getInternalName(competitor.name);
-    const stockPrice = Helpers.CalculateStockPrice(competitor, Helpers.CalculateValuation(competitor));
-
-    GetRootScope().settings.competitorProducts.push({
-        id: internalName,
-        name: competitor.name,
-        logoPath: competitor.logoPath,
-        logoColorDegree: competitor.logoColorDegree,
-        users: competitor.users,
-        productTypeName: competitor.productTypeName,
-        version: 100,
-        controlled: false,
-        history: [{
-            day: 1,
-            stockPrice: stockPrice,
-            users: competitor.users,
-            week: 1,
-        }],
-        stockVolume: competitor.stockVolume,
-        ownedStocks: 0,
-        dealResults: [],
-        stockTransactions: [],
-        priceExpectations: Math.round((Math.random() * 3 + 2) * 10) / 10,
-    });
-}
-
-function registerFeature(feature) {
-    const internalName = getInternalName(feature.name);
-
-    if(FeatureNames[internalName]) {
-        return;
-    }
-
-    // Add feature
-    FeatureNames[internalName] = internalName;
-    const clone = Helpers.Clone(feature);
-    clone.name = internalName;
-    Features.push(clone);
-
-    // Add research items
-    ResearchItemNames[internalName] = internalName;
-    ResearchItems.push({
-        name: internalName,
-        category: 'Features',
-        unlockType: 'Feature',
-        points: feature.researchPoints,
-    });
-
-    // Add feature to specified product types
-    feature.availableProducts.map(name => ProductTypes.find(product => product.name == name)).forEach(product => {
-        if(product) {
-            product.features.push(internalName);
-        }
-    });
-
-    // Add language strings
-    Modding.addTranslation(internalName, {
-        en: feature.name
-    });
-    // TODO: Add option for description
-    Modding.addTranslation(internalName + '_description', {
-        en: 'The multipurpose feature ' + feature.name + ' enables every animal, human, robot and plant to live together in harmony. <3'
-    });
-}
-
-function registerFramework(framework) {
-    const internalName = getInternalName(framework.name);
-
-    if(FrameworkNames[internalName]) {
-        return;
-    }
-
-    // Add framework
-    FrameworkNames[internalName] = internalName;
-
-    const clone = Helpers.Clone(framework);
-    clone.name = internalName;
-    clone.logoPath = _modPath + 'thumbnail.png';
-    clone.order = _.maxBy(Frameworks, e => e.order).order + 1;
-    delete clone.researchPoints;
-    Frameworks.push(clone);
-
-    // Add research items
-    ResearchItemNames[internalName] = internalName;
-    ResearchItems.push({
-        name: internalName,
-        category: 'Frameworks',
-        unlockType: 'Framework',
-        points: framework.researchPoints,
-    });
-
-    // Add language strings
-    Modding.addTranslation(internalName, {
-        en: framework.name,
-    });
-}
-
-function registerProduct(product) {
-    const internalName = getInternalName(product.name);
-
-    if(ProductTypeNames[internalName]) {
-        console.log('Skipping product', product, 'because it already exists');
-        return;
-    }
-
-    ProductTypeNames[internalName] = internalName;
-    const productCopy = JSON.parse(JSON.stringify(product));
-    productCopy.name = internalName
-    ProductTypes.push(productCopy);
-    Modding.addTranslation(internalName, {
-        en: product.name
-    });
-}
